#' @export createMetaDisco <- function(condition1, condition2, n_tile, window_size, k_neigh, bandwidth, n_bootstrap, bootstrap_method, weight_scheme, n_cores = parallel::detectCores()) list1 <- list() for(i in 1:length(condition1)) print(condition1i) if(is.list(condition1)) list1[i] <- createSpomic(condition1[i]) else list1[i] <- createSpomic(condition1i) # list1[i] <- createSpomic(condition1i) list1[i] <- setSpomicHypers(spomic_obj=list1[i], n_tile=n_tile, window_size=window_size, k_neigh=k_neigh, bandwidth=bandwidth, n_bootstrap=n_bootstrap, weight_scheme=weight_scheme) list1[i] <- runCLQ(spomic_obj=list1[i], bootstrap_method=bootstrap_method, n_cores=n_cores) createMetaDisco( condition1, condition2, tile_size, window_size, k_neigh, bandwidth, n_bootstrap, weight_scheme, n_cores = parallel::detectCores() ) condition1A list of Spomic objects, data frames of paths to CSV files for the second condition. tile_sizeA numeric value specifying the size of tiles to be used for spatial bootstrapping. window_sizeAn integer specifying the size of the spatial window for analysis. k_neighAn integer representing the number of nearest neighbors to consider in colocalization analysis. bandwidthAn integer defining the bandwidth used to restrict colocalization analysis. n_bootstrapAn integer specifying the number of bootstrapping iterations to perform. weight_schemeA character string describing the weighting scheme for colocalization. n_coresAn integer specifying the number of CPU cores to use for parallel processing. Defaults to the number of available cores. A MetaDisco object containing processed spatial omics data for both conditions, along with the analysis parameters. Constructs a MetaDisco object from two conditions, each containing one or more Spomic objects or data respresentations that can be converted into Spomic objects. This function processes spatial omics data for both conditions, sets analysis parameters (hypers), and performs colocalization analysis with optional bootstrapping.